WIP -- Experimental GIL work for Wallaroo Python
I've been able to get the word count demo to start and to accept a few
messages, but it looks like it tries to acquire the lock in the state
computation and get then hangs. This happens with a single thread or 4
threads, and happens whether I'm running a single worker or multiple
workers. I'm not sure if the lock didn't get released somewhere
earlier, or if there's an issue somewhere else.

In order to compile this version of Machida you will need to add the
following things to the pony compiler:

src/libponyrt/actor/actor.c

+PONY_API int32_t pony_scheduler_index(pony_ctx_t* ctx)
+{
+  return ctx->scheduler->index;
+}

src/libponyrt/pony.h

+PONY_API int32_t pony_scheduler_index(pony_ctx_t* ctx);