Step 9: Ooops, `Bus error` failure, cool!
If we compile with `-d` debug mode, then when we also change
the maximum number of claimants with this patch from 50 -> 250,
then we can cause the entire program to crash with `Bus error`.

The `_coalesce()` algorithm is very naive and can use a large
amount of stack space.  Exhausting the scheduler's Pthread
stack space is the likely cause of the crash.

% ( ponyc -d --verbose=1 . ) && ./hash_partitions --sequential
Generating
Writing ./hash_partitions.o
Linking ./hash_partitions
1 test started, 0 complete: hash_partitions/ponycheck started
.Bus error