WIP: racy DOSclient actor
Hrrrrm, I thought that the TCP code will buffer data
that is sent before connection.  But that doesn't
appear to be the case.  (Perhaps it's the Wallaroo
forks that do the buffering but stdlib does not?)
AAAAhhhhhh, correct, stdlib's write() and writev()
both start with `if _connected and...`

When the @sleep is uncommented to avoid the race:

    SOCK: I am connected.
    DOSc: send_ls
    SOCK: sent
    SOCK: received header
    SOCK: received payload
    DOSclient GOT:bar	no
    baz	no
    foo	no

Without the sleep:

    DOSc: send_ls
    SOCK: I am connected.