Make inputs and outputs explicit (and store by id)
Previously we only kept track of the inputs
to a Consumer as a set of upstream Producers and outputs
from a Producer implicitly in the Router. This commit adds
bookkeeping such that inputs and outputs are kept track of
explicitly by StepId. This means that we can now distinguish
multiple inputs coming from the same DataReceiver and multiple
outputs over the same OutgoingBoundary.