Implement checkpoint barrier infrastructure
In order to checkpoint for later recovery, we need to inject barriers into
the datastream that are used to partition data into checkpoint and post-
checkpoint messages. The checkpoint barriers are injected at all sources
and flow downstream through the chain of Consumers. Each Consumer needs to
block an input once it has received a barrier on that input. Once it receives
a barrier on all inputs, it can resume normal processing.