Implement topology graph rollback
Rollback consists of two aspects: (1) We need to rollback to the earlier
topology graph and (2) We need to rollback individual producers and consumers
to their state during the checkpoint we're rolling back to. This commit
implements (1). Topology graph rollback requires that we determine which
local keys we were responsible for during the target checkpoint and then doing
three things: (a) If we no longer have that key, create a state step for it,
(b) if we have a key that we did not have, dispose the corresponding state
step and remove it from the graph, and (c) if we still have the key, do
nothing. The actual state rollback on each state step is part of (2).